Split setup help parsing from file planning
